requirements.txt
glob2==0.7
tqdm==4.62.3
librosa==0.8.1
numba==0.56.3
numpy
scipy==1.10.1
tensorboard
torch
torchvision
torchaudio
pandas==2.0.3
plotly
praatio<=5.0.0
praat-parselmouth>=0.4.0
tgt>=1.4.0
pandas>=1.4.0
soundfile>=0.10.0
PyYAML==6.0.2
rich>=13.2.0
scikit-learn>=1.2.0
ruamel.yaml==0.17.21
webrtcvad==2.0.10
nbformat>=4.2.0
whisperx